pyproject.toml
[project]
name = "debugger"
version = "0.1.0"
description = "Contract-driven multi-track debugging + hardening orchestrator"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "typer>=0.12.3",
  "rich>=13.7.1",
  "pydantic>=2.8.2",
  "pyyaml>=6.0.2",
  "jsonschema>=4.23.0",
  "loguru>=0.7.2",
]

[project.optional-dependencies]
treesitter = [
  "tree_sitter>=0.22.3",
  "tree_sitter_languages>=1.10.2",
]

[dependency-groups]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
]
ast = [
  "treesitter-tools @ file:///home/graham/workspace/experiments/treesitter-tools",
]

[project.scripts]
dbg = "debugger.cli:app"

[build-system]
requires = ["hatchling>=1.25.0"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/debugger"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP"]
